requirements.txt
numpy==1.26.4
pandas==2.2.2
python-dateutil==2.9.0.post0
python-poppler==0.4.1
pytz==2024.1
six==1.16.0
tzdata==2024.1
